+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Replaced Super Sonic "Hmph" with "I'll show you what the Chaos Emeralds can really do!" in the character select
   screen.
+- Removed 10 second penalty at the end of Sonic and Tails levels.
 
 ## [v0.3.1] - 2024-07-23